refactor: extract PDF conversion reporting
